refactor(petlist): separate client and server components for Vercel deployment
